pkg/function-chat/requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
requests==2.30.0
pydantic==1.10.9
uuid==1.30
openai==0.27.8
PyYAML==6.0.1